Rename odometry to dead-reconning, create odometry (#12)
implements #11, #7,  and component part of bmegli/ev3dev-mapping#10

Implements bmegli/ev3dev-mapping#8 as Physics componenet


Also simplify class hierarchy.

Got rid off:
- MotionModels class hierarchy
- DifferentialDrive class hierarchy

It's not necessary to complicate at this point. This can be reintrocuded
as needed.

For now Odometry has binary compatiblity with DeadReconnning so that
replays can be shared to rebuild with odoemtry model from dead reconning
record. This is not guaranteed to be compatible in the future.